converted C.int to int, and C.double to float64, to make it normal more like a native go functions

+++ funcs.go
@@ -15,9 +15,9 @@ return Zero'd Hdate_Struct
 */
 func Init() *Hdate_Struct {
 	var h *Hdate_Struct = new(Hdate_Struct)
-  var d *C.hdate_struct
-  d = C.new_hdate()
-  h.d = d
+	var d *C.hdate_struct
+	d = C.new_hdate()
+	h.d = d
 
 	return h
 }
@@ -26,7 +26,7 @@ func Init() *Hdate_Struct {
 Destruct memory of Hdate_Struct, must be called last
 */
 func (h *Hdate_Struct) Destruct() {
-  C.delete_hdate(h.d)
+	C.delete_hdate(h.d)
 }
 
 /**
@@ -38,8 +38,8 @@ m Month 1..12
 y Year in 4 digits e.g. 2001
 */
 
-func (h *Hdate_Struct) Set_Gdate(d, m, y C.int) {
-	hdate := C.hdate_set_gdate(h.d, d, m, y)
+func (h *Hdate_Struct) Set_Gdate(d, m, y int) {
+	hdate := C.hdate_set_gdate(h.d, C.int(d), C.int(m), C.int(y))
 	if hdate != h.d {
 		h.d = hdate
 	}
@@ -53,8 +53,8 @@ m Month 1..14 ,(13 - Adar 1, 14 - Adar 2)
 	if m or d is 0 return current date.
 y Year in 4 digits e.g. 5731
 */
-func (h *Hdate_Struct) Set_Hdate(d, m, y C.int) {
-	hdate := C.hdate_set_hdate(h.d, d, m, y)
+func (h *Hdate_Struct) Set_Hdate(d, m, y int) {
+	hdate := C.hdate_set_hdate(h.d, C.int(d), C.int(m), C.int(y))
 	if hdate != h.d {
 		h.d = hdate
 	}
@@ -65,8 +65,8 @@ compute date structure from the Julian day
 
 jd the julian day number.
 */
-func (h *Hdate_Struct) Set_jd(jd C.int) {
-	hdate := C.hdate_set_jd(h.d, jd)
+func (h *Hdate_Struct) Set_jd(jd int) {
+	hdate := C.hdate_set_jd(h.d, C.int(jd))
 	if hdate != h.d {
 		h.d = hdate
 	}
@@ -86,11 +86,11 @@ warning This was originally written using a local static string,
          calling for output to be copied away.
 
 */
-func (h *Hdate_Struct) Get_Format_Date(diaspora, s C.int) string {
+func (h *Hdate_Struct) Get_Format_Date(diaspora, s int) string {
 	var ch *C.char
 	str := ""
 
-	ch = C.hdate_get_format_date(h.d, diaspora, s)
+	ch = C.hdate_get_format_date(h.d, C.int(diaspora), C.int(s))
 	if ch != nil {
 		str = C.GoString(ch)
 		C.free(unsafe.Pointer(ch))
@@ -106,8 +106,8 @@ param diaspora if true give diaspora readings
 return the number of parasha 1. Bereshit etc..
   (55 through 61 are joined strings e.g. Vayakhel Pekudei)
 */
-func (h *Hdate_Struct) Get_Parasha(diaspora C.int) C.int {
-	return C.hdate_get_parasha(h.d, diaspora)
+func (h *Hdate_Struct) Get_Parasha(diaspora int) int {
+	return int(C.hdate_get_parasha(h.d, C.int(diaspora)))
 }
 
 /**
@@ -116,8 +116,8 @@ get the number of hebrew holiday.
 diaspora if true give diaspora holidays
 return the number of holiday.
 */
-func (h *Hdate_Struct) Get_Holyday(diaspora C.int) C.int {
-	return C.hdate_get_holyday(h.d, diaspora)
+func (h *Hdate_Struct) Get_Holyday(diaspora int) int {
+	return int(C.hdate_get_holyday(h.d, C.int(diaspora)))
 }
 
 /**
@@ -126,8 +126,8 @@ Return the day in the omer of the given date
 param h The hdate_struct of the date to use.
 return The day in the omer, starting from 1 (or 0 if not in sfirat ha omer)
 */
-func (h *Hdate_Struct) Get_Omer_Day() C.int {
-	return C.hdate_get_omer_day(h.d)
+func (h *Hdate_Struct) Get_Omer_Day() int {
+	return int(C.hdate_get_omer_day(h.d))
 }
 
 /**
@@ -138,8 +138,8 @@ Holiday types: set as constants as HOLYDAY_XXXX
 param holyday the holyday number
 return the number of holyday type.
 */
-func Get_Holyday_Type(holyday C.int) C.int {
-	return C.hdate_get_holyday_type(holyday)
+func Get_Holyday_Type(holyday int) int {
+	return int(C.hdate_get_holyday_type(C.int(holyday)))
 }
 
 /**
@@ -148,8 +148,8 @@ size of hebrew year in days.
 param hebrew_year the hebrew year.
 return size of Hebrew year
 */
-func Get_Hebrew_Year_Size(year C.int) C.int {
-	return C.hdate_get_size_of_hebrew_year(year)
+func Get_Hebrew_Year_Size(year int) int {
+	return int(C.hdate_get_size_of_hebrew_year(C.int(year)))
 }
 
 /**
@@ -160,8 +160,8 @@ author Amos Shapir 1984 (rev. 1985, 1992) Yaacov Zamir 2003-2005
 param hebrew_year The Hebrew year
 return Number of days since 3,1,3744
 */
-func Days_from_3744(year C.int) C.int {
-	return C.hdate_days_from_3744(year)
+func Days_from_3744(year int) int {
+	return int(C.hdate_days_from_3744(C.int(year)))
 }
 
 /**
@@ -171,8 +171,8 @@ param size_of_year Length of year in days
 param new_year_dw First week day of year
 return the number for year type (1..14)
 */
-func Get_Year_Type(year_size, new_year_dw C.int) C.int {
-	return C.hdate_get_year_type(year_size, new_year_dw)
+func Get_Year_Type(year_size, new_year_dw int) int {
+	return int(C.hdate_get_year_type(C.int(year_size), C.int(new_year_dw)))
 }
 
 /**
@@ -185,8 +185,8 @@ param month Month 1..12
 param year Year in 4 digits e.g. 2001
 return the julian day number
 */
-func GDate_to_JD(day, month, year C.int) C.int {
-	return C.hdate_gdate_to_jd(day, month, year)
+func GDate_to_JD(day, month, year int) int {
+	return int(C.hdate_gdate_to_jd(C.int(day), C.int(month), C.int(year)))
 }
 
 /**
@@ -205,16 +205,17 @@ param jd_tishrey1 return the julian number of 1 Tishrey this year
 param jd_tishrey1_next_year return the julian number of 1 Tishrey next year
 return the julian day number
 */
-func HDate_To_JD(day, month, year C.int) Hdate_Julian {
+func HDate_To_JD(day, month, year int) Hdate_Julian {
 	var jd_tishrey1 C.int
 	var jd_tishrey1_next_year C.int
 
-	result := C.hdate_hdate_to_jd(day, month, year,
+	result := C.hdate_hdate_to_jd(C.int(day), C.int(month), C.int(year),
 		&jd_tishrey1, &jd_tishrey1_next_year)
 
-	return Hdate_Julian{JD_Tishrey1: jd_tishrey1,
-		JD_Tishrey1_Next_Year: jd_tishrey1_next_year,
-		Day: result,
+	return Hdate_Julian{
+		JD_Tishrey1:           int(jd_tishrey1),
+		JD_Tishrey1_Next_Year: int(jd_tishrey1_next_year),
+		Day: int(result),
 	}
 }
 
@@ -226,8 +227,8 @@ param month this month
 param year this year
 return the days from 1 jan
 */
-func Get_Day_of_Year(day, month, year C.int) C.int {
-	return C.hdate_get_day_of_year(day, month, year)
+func Get_Day_of_Year(day, month, year int) int {
+	return int(C.hdate_get_day_of_year(C.int(day), C.int(month), C.int(year)))
 }
 
 /**
@@ -252,13 +253,14 @@ return:
 param sunrise return the utc sunrise in minutes
 param sunset return the utc sunset in minutes
 */
-func Get_UTC_Sun_Time_Deg(day, month, year C.int, latitude, longitude, deg C.double) Sunrise_Sunset {
+func Get_UTC_Sun_Time_Deg(day, month, year int, latitude, longitude, deg float64) Sunrise_Sunset {
 	var sunrise, sunset C.int
-	C.hdate_get_utc_sun_time_deg(day, month, year, latitude, longitude, deg, &sunrise, &sunset)
+	C.hdate_get_utc_sun_time_deg(C.int(day), C.int(month), C.int(year),
+		C.double(latitude), C.double(longitude), C.double(deg), &sunrise, &sunset)
 
 	return Sunrise_Sunset{
-		Sunrise: sunrise,
-		Sunset:  sunset,
+		Sunrise: int(sunrise),
+		Sunset:  int(sunset),
 	}
 }
 
@@ -277,13 +279,14 @@ returns
 param sunrise return the utc sunrise in minutes after midnight (00:00)
 param sunset return the utc sunset in minutes after midnight (00:00)
 */
-func Get_UTC_Sun_Time(day, month, year C.int, latitude, longitude C.double) Sunrise_Sunset {
+func Get_UTC_Sun_Time(day, month, year int, latitude, longitude float64) Sunrise_Sunset {
 	var sunrise, sunset C.int
-	C.hdate_get_utc_sun_time(day, month, year, latitude, longitude, &sunrise, &sunset)
+	C.hdate_get_utc_sun_time(C.int(day), C.int(month), C.int(year),
+		C.double(latitude), C.double(longitude), &sunrise, &sunset)
 
 	return Sunrise_Sunset{
-		Sunrise: sunrise,
-		Sunset:  sunset,
+		Sunrise: int(sunrise),
+		Sunset:  int(sunset),
 	}
 }
 
@@ -306,23 +309,24 @@ param sunset return the utc sunset in minutes
 param first_stars return the utc tzeit hacochavim in minutes
 param three_stars return the utc shlosha cochavim in minutes
 */
-func Get_UTC_Sun_Time_Full(day, month, year C.int, latitude, longitude C.double) Sun_Time {
+func Get_UTC_Sun_Time_Full(day, month, year int, latitude, longitude float64) Sun_Time {
 	var sun_hour, first_light, talit, sunrise, midday, sunset C.int
 	var first_stars, three_stars C.int
 
-	C.hdate_get_utc_sun_time_full(day, month, year, latitude, longitude,
+	C.hdate_get_utc_sun_time_full(C.int(day), C.int(month), C.int(year),
+		C.double(latitude), C.double(longitude),
 		&sun_hour, &first_light, &talit, &sunrise, &midday, &sunset,
 		&first_stars, &three_stars)
 
 	return Sun_Time{
-		Sun_Hour:    sun_hour,
-		First_Light: first_light,
-		Talit:       talit,
-		Sunrise:     sunrise,
-		MidDay:      midday,
-		Sunset:      sunset,
-		First_Stars: first_stars,
-		Three_Stars: three_stars,
+		Sun_Hour:    int(sun_hour),
+		First_Light: int(first_light),
+		Talit:       int(talit),
+		Sunrise:     int(sunrise),
+		MidDay:      int(midday),
+		Sunset:      int(sunset),
+		First_Stars: int(first_stars),
+		Three_Stars: int(three_stars),
 	}
 }
 
@@ -331,8 +335,8 @@ get the Gregorian day of the month
 
 return the Gregorian day of the month, 1..31.
 */
-func (h *Hdate_Struct) Get_GDay() C.int {
-	return C.hdate_get_gday(h.d)
+func (h *Hdate_Struct) Get_GDay() int {
+	return int(C.hdate_get_gday(h.d))
 }
 
 /**
@@ -340,8 +344,8 @@ get the Gregorian month
 
 return the Gregorian month, jan = 1.
 */
-func (h *Hdate_Struct) Get_GMonth() C.int {
-	return C.hdate_get_gmonth(h.d)
+func (h *Hdate_Struct) Get_GMonth() int {
+	return int(C.hdate_get_gmonth(h.d))
 }
 
 /**
@@ -349,8 +353,8 @@ get the Gregorian year
 
 return the Gregorian year.
 */
-func (h *Hdate_Struct) Get_GYear() C.int {
-	return C.hdate_get_gyear(h.d)
+func (h *Hdate_Struct) Get_GYear() int {
+	return int(C.hdate_get_gyear(h.d))
 }
 
 /**
@@ -358,8 +362,8 @@ get the Hebrew day of the month
 
 return the Hebrew day of the month, 1..30.
 */
-func (h *Hdate_Struct) Get_HDay() C.int {
-	return C.hdate_get_hday(h.d)
+func (h *Hdate_Struct) Get_HDay() int {
+	return int(C.hdate_get_hday(h.d))
 }
 
 /**
@@ -367,8 +371,8 @@ get the Hebrew month
 
 return the Hebrew month, Tishery = 1 .. Adar I =13, Adar II = 14.
 */
-func (h *Hdate_Struct) Get_HMonth() C.int {
-	return C.hdate_get_hmonth(h.d)
+func (h *Hdate_Struct) Get_HMonth() int {
+	return int(C.hdate_get_hmonth(h.d))
 }
 
 /**
@@ -376,8 +380,8 @@ get the Hebrew year
 
 return the Hebrew year.
 */
-func (h *Hdate_Struct) Get_HYear() C.int {
-	return C.hdate_get_hyear(h.d)
+func (h *Hdate_Struct) Get_HYear() int {
+	return int(C.hdate_get_hyear(h.d))
 }
 
 /**
@@ -385,8 +389,8 @@ get the day of the week
 
 return the the day of the week.
 */
-func (h *Hdate_Struct) Get_Day_Of_Week() C.int {
-	return C.hdate_get_day_of_the_week(h.d)
+func (h *Hdate_Struct) Get_Day_Of_Week() int {
+	return int(C.hdate_get_day_of_the_week(h.d))
 }
 
 /**
@@ -394,8 +398,8 @@ get the size of the hebrew year
 
 return the the size of the hebrew year.
 */
-func (h *Hdate_Struct) Get_Size_of_Year() C.int {
-	return C.hdate_get_size_of_year(h.d)
+func (h *Hdate_Struct) Get_Size_of_Year() int {
+	return int(C.hdate_get_size_of_year(h.d))
 }
 
 /**
@@ -403,8 +407,8 @@ get the new year day of the week
 
 return the the new year day of the week.
 */
-func (h *Hdate_Struct) Get_New_Year_Day_of_Week() C.int {
-	return C.hdate_get_new_year_day_of_the_week(h.d)
+func (h *Hdate_Struct) Get_New_Year_Day_of_Week() int {
+	return int(C.hdate_get_new_year_day_of_the_week(h.d))
 }
 
 /**
@@ -412,8 +416,8 @@ get the Julian day number
 
 return the Julian day number.
 */
-func (h *Hdate_Struct) Get_Julian() C.int {
-	return C.hdate_get_julian(h.d)
+func (h *Hdate_Struct) Get_Julian() int {
+	return int(C.hdate_get_julian(h.d))
 }
 
 /**
@@ -421,8 +425,8 @@ get the number of days passed since 1 tishrey
 
 return the number of days passed since 1 tishrey.
 */
-func (h *Hdate_Struct) Get_Days() C.int {
-	return C.hdate_get_days(h.d)
+func (h *Hdate_Struct) Get_Days() int {
+	return int(C.hdate_get_days(h.d))
 }
 
 /**
@@ -430,8 +434,8 @@ get the number of weeks passed since 1 tishrey
 
 return the number of weeks passed since 1 tishrey.
 */
-func (h *Hdate_Struct) Get_Weeks() C.int {
-	return C.hdate_get_weeks(h.d)
+func (h *Hdate_Struct) Get_Weeks() int {
+	return int(C.hdate_get_weeks(h.d))
 }
 
 /**
@@ -487,8 +491,11 @@ func Is_Hebrew_Locale() bool {
  return  a string containing the information.
          return empty string upon failure.
 */
-func Hdate_String(string_type, index, short_form, hebrew_form C.int) string {
-	ch := C.hdate_string(string_type, index, short_form, hebrew_form)
+func Hdate_String(string_type, index, short_form, hebrew_form int) string {
+	ch := C.hdate_string(C.int(string_type),
+		C.int(index),
+		C.int(short_form),
+		C.int(hebrew_form))
 	result := ""
 
 	if ch != nil {
